+++ app/api/inventory/stockout/route.ts
@@ -9,6 +9,11 @@ export async function GET(request: NextRequest) {
       quantityOnHand: {
         gt: 0,
       },
+      grade: {
+        description: {
+          not: "Ungraded",
+        },
+      },
     },
     select: {
       grade: true,
