+++ components/index.ts
@@ -10,6 +10,9 @@ import Pagination from "./dashboard/Pagination";
 import GradePage from "./dashboard/GradePage";
 import ForgotPasswordModal from "./auth/ForgotPasswordModal";
 import StockOutModal from "./inventory/stockout/StockOutModal";
+import InventoryUpdateForm from "./inventory/inventory/InventoryUpdateForm";
+import SortModal from "./inventory/inventory/SortModal";
+import StockoutModal from "./inventory/inventory/StockoutModal";
 export {
   Sidepanel,
   InventoryInputForm,
@@ -22,4 +25,7 @@ export {
   GradePage,
   ForgotPasswordModal,
   StockOutModal,
+  InventoryUpdateForm,
+  SortModal,
+  StockoutModal,
 };
