+++ components/inventory/inventory/StockoutModal.tsx
@@ -1,0 +1,13 @@
+import Swal from "sweetalert2";
+import { inventoryDataType } from "@/types";
+const StockoutModal = ({
+  swal,
+  inventory,
+}: {
+  swal: typeof Swal;
+  inventory: inventoryDataType;
+}) => {
+  return <div>StockoutModal</div>;
+};
+
+export default StockoutModal;
